fix: countTotalAndWonBattles 타입 변경

See merge request s11-bigdata-recom-sub1/S11P21C106!74

+++ project/back/health/src/main/java/com/ssafy/health/domain/crew/dto/response/CrewDetailResponseDto.java
@@ -13,8 +13,8 @@ public class CrewDetailResponseDto {
     private String profileImage;
     private Integer crewCoins;
     private Long crewRanking;
-    private Integer totalBattlesCount;
-    private Integer winCount;
+    private Long totalBattlesCount;
+    private Long winCount;
     private Float averageAge;
     private Integer basicScore;
     private Integer activityScore;
